many things and optimizations, plus framebuffers dont work lol

+++ src/main/kotlin/cc/polyfrost/polyui/renderer/data/Framebuffer.kt
@@ -2,7 +2,7 @@ package cc.polyfrost.polyui.renderer.data
 
 import cc.polyfrost.polyui.properties.Settings
 
-data class Framebuffer(val width: Int, val height: Int, val addr: Int, val type: Settings.BufferType) {
+data class Framebuffer(val width: Float, val height: Float, val addr: Int, val type: Settings.BufferType) {
 
     enum class Mode {
         Read, Write, ReadWrite
